+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 0.2.0 (in development)
 
 * Inclusion of predicted resistances to antimicrobial drugs thanks to gene/drug mappings from the NARMS/CIPARS Molecular Working Group. Resistance predictions are microbiological resistances and not clinical resistances. 
+* Switched to including negative AMR results (samples with no AMR genes) by default.  Must now use parameter `--exclude-negatives` to exclude them.
 
 # Version 0.1.0